- 解决 mainwindow.ui / mainwindow.cpp 合并冲突，统一采用代码版 UI
- 将 res.qrc 加入 CMake 构建，删除重复且路径错误的 assets/res.qrc，使 SVG 地图资源能被编译进程序
- 移除 mainwindow.h 中未使用的 Ui::MainWindow 成员与前向声明
- 修复 Windows 下 long(32位) 装不下世界人口 8059506979 的溢出问题（改用 long long）

+++ mainwindow.cpp
@@ -214,8 +214,8 @@ void MainWindow::updateDayUI(int day)
 
 void MainWindow::updateStatsUI(long totalInfected, long totalDead)
 {
-    long totalPopulation = 8059506979; // 假定总人口
-    long healthy = totalPopulation - totalInfected - totalDead;
+    long long totalPopulation = 8059506979LL; // 假定总人口（注意：Windows 上 long 为 32 位，需用 long long）
+    long long healthy = totalPopulation - totalInfected - totalDead;
 
     QLabel *worldDataLabel = this->findChild<QLabel*>("WorldDataLabel");
     if (worldDataLabel) {
